DAOS-2721 control: Reorganize binaries under cmd/
When a Go project produces multiple binaries, it is a common
convention to place each binary's main.go in a subdirectory
of cmd/ named after the produced binary. e.g. cmd/daos_server

This organization convention allows for easy identification of
which binaries could be produced from the project, and takes
advantage of a property of the Go compiler which automatically
produces a binary named after the directory where main.go is
located.
