+++ src/control/SConscript
@@ -33,7 +33,7 @@ def cmd_build(gopath, repopath, cmd):
 
 def get_install_src_dir(repopath, name):
     """Get the Gopath-based directory to run 'go install' on"""
-    return join(repopath, "src", "control", name)
+    return join(repopath, "src", "control", "cmd", name)
 
 def get_install_bin_path(env, name):
     """Get the final installation location for a binary with a given name"""
@@ -62,7 +62,7 @@ def install_go_bin(denv, gosrc, gopath, libs, name, install_name):
     path = join(gopath, "src", repopath, "src")
 
     bin_path = get_bin_path(gopath, name)
-    mod_src = join(gosrc, name, "main.go") # Module src
+    mod_src = join(gosrc, "cmd", name, "main.go") # Module src
     install_src = get_install_src_dir(repopath, name)
     install_bin = get_install_bin_path(denv, install_name)
     src = [mod_src]
@@ -152,12 +152,12 @@ def scons():
                 Copy("$TARGET", "$SOURCE")
             ])
 
-    install_go_bin(denv, gosrc, gopath, [nvmecontrol], "server", "daos_server")
+    install_go_bin(denv, gosrc, gopath, [nvmecontrol], "daos_server", "daos_server")
     install_go_bin(denv, gosrc, gopath, None, "dmg", "daos_shell")
     install_go_bin(denv, gosrc, gopath, None, "drpc_test", "hello_drpc")
 
     agentbin = get_bin_path(gopath, "agent")
-    serverbin = get_bin_path(gopath, "server")
+    serverbin = get_bin_path(gopath, "daos_server")
     shellbin = get_bin_path(gopath, "dmg")
     drpcbin = get_bin_path(gopath, "drpc_test")
     AlwaysBuild([agentbin, serverbin, shellbin, drpcbin])
