+++ src/control/cmd/agent/mgmt_rpc.go
@@ -23,7 +23,7 @@
 
 package main
 
-// #cgo CFLAGS: -I${SRCDIR}/../../include
+// #cgo CFLAGS: -I${SRCDIR}/../../../include
 // #include <daos/drpc_modules.h>
 import "C"
