+++ src/control/cmd/drpc_test/main.go
@@ -30,8 +30,8 @@ import (
 	"os/signal"
 	"syscall"
 
+	"github.com/daos-stack/daos/src/control/cmd/drpc_test/hello"
 	"github.com/daos-stack/daos/src/control/drpc"
-	"github.com/daos-stack/daos/src/control/drpc_test/hello"
 	"github.com/daos-stack/daos/src/control/log"
 	"github.com/golang/protobuf/proto"
 	"github.com/pkg/errors"
