+++ src/control/server/mgmt_drpc.go
@@ -21,9 +21,9 @@
 // portions thereof marked with this legend must also reproduce the markings.
 //
 
-package main
+package server
 
-// #cgo CFLAGS: -I${SRCDIR}/../../include
+// #cgo CFLAGS: -I${SRCDIR}/../../../include
 // #include <daos/drpc_modules.h>
 import "C"
