+++ src/control/server/mgmt_storage.go
@@ -21,7 +21,7 @@
 // portions thereof marked with this legend must also reproduce the markings.
 //
 
-package main
+package server
 
 import (
 	"github.com/daos-stack/daos/src/control/common"
